refactor cards
