+++ frontend/components/LoginPage/cards/Card3/index.tsx
@@ -1,31 +1,31 @@
-import darkImage from "./question-dark.jpg";
-import lightImage from "./question-light.jpg";
+import darkImage from "./matching-dark.jpg";
+import lightImage from "./matching-light.jpg";
 import TemplateCard, { ContentPosition } from "../TemplateCard";
 import { Button } from "@nextui-org/react";
-import { scrollTillCallToAction } from "../utils";
+import { scrollTillCallToAction } from "../../utils";
 
-const Card2 = () => {
+const Card3 = () => {
   return (
     <TemplateCard
-      contentPosition={ContentPosition.RIGHT}
+      contentPosition={ContentPosition.LEFT}
       content={
         <div>
-          <div>Attempt the latest LeetCode questions</div>
+          <div>Collaborate with peers</div>
           <div className="flex justify-end w-full mt-5">
             <Button
               color="secondary"
               variant="bordered"
               onClick={scrollTillCallToAction}
             >
-              Begin now!
+              Find a peer!
             </Button>
           </div>
         </div>
       }
-      contentClassName="mt-[45px]"
-      imageProps={{ darkImage, lightImage, alt: "login-card-2" }}
+      contentClassName="mt-[15px]"
+      imageProps={{ darkImage, lightImage, alt: "login-card-3" }}
     />
   );
 };
 
-export default Card2;
+export default Card3;
