+++ frontend/components/LoginPage/cards/TemplateCard/ImageCard/index.tsx
