+++ frontend/components/LoginPage/index.tsx
@@ -1,8 +1,9 @@
-import Card1 from "./Card1";
-import Card2 from "./Card2";
-import Card3 from "./Card3";
-import Card4 from "./Card4";
-import LoginButtonGroup, { Providers } from "./LoginButtonGroup";
+import CallToActionCard from "./cards/CallToActionCard";
+import Card1 from "./cards/Card1";
+import Card2 from "./cards/Card2";
+import Card3 from "./cards/Card3";
+import Card4 from "./cards/Card4";
+import { Providers } from "./LoginButtonGroup";
 
 type LoginPageComponentProps = {
   providers: Providers;
@@ -17,12 +18,7 @@ const LoginPageComponent = (props: LoginPageComponentProps) => {
       <Card2 />
       <Card3 />
       <Card4 />
-      <div className="w-full h-[200px] mb-40 text-center">
-        <div id={CALL_TO_ACTION} className="text-[3em] font-semibold mb-3">
-          Start coding!
-        </div>
-        <LoginButtonGroup {...props} />
-      </div>
+      <CallToActionCard {...props} />
     </div>
   );
 };
